add bower
