+++ Gruntfile.js
@@ -1,0 +1,14 @@
+/**
+ * <project_name>
+ * <project_repo>
+ *
+ * Copyright (c) <year> <username>
+ * Licensed under the MIT license.
+ *
+ */
+/*jslint indent: 4 */
+/*global module, require */
+module.exports = function (grunt) {
+    'use strict';
+    require('load-grunt-config')(grunt);
+};
