+++ grunt/aliases.yml
@@ -1,0 +1,6 @@
+build:
+    - clean
+    - copy
+    - uglify
+default:
+    - build
