+++ grunt/clean.js
@@ -1,0 +1,8 @@
+/*jslint indent: 4 */
+/*global module */
+module.exports = function (grunt, opts) {
+    'use strict';
+    return {
+        build: ['dist/**/*'],
+    };
+};
