+++ package.json
@@ -1,40 +1,46 @@
 {
-    "name": "isuri",
-    "description": "Pure Javascript implementation for truly checking if the provided input is an URI. Based on RFC 3986.",
-    "version": "1.1.0",
-    "author": {
-        "name" : "David Pate",
-        "email" : "me@davidtpate.com",
-        "url" : "http://davidtpate.com"
-    },
-    "keywords": [
-        "uri",
-        "url",
-        "type",
-        "checker",
-        "validator",
-        "validate"
-    ],
-    "homepage": "https://github.com/DavidTPate/isuri",
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/DavidTPate/isuri.git"
-    },
-    "bugs": {
-        "url": "https://github.com/DavidTPate/isuri/issues"
-    },
-    "license": "MIT",
-    "main": "index.js",
-    "devDependencies": {
-        "istanbul": "^0.x",
-        "mocha": "^1.x",
-        "should": "^4.x",
-        "benchmark": "^1.x",
-        "beautify-benchmark": "^0.x"
-    },
-    "scripts": {
-        "test": "mocha --check-leaks --bail --reporter spec test/",
-        "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --check-leaks --reporter dot test/",
-        "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --check-leaks --reporter spec test/"
-    }
+  "name": "isuri",
+  "description": "Pure Javascript implementation for truly checking if the provided input is an URI. Based on RFC 3986.",
+  "version": "1.1.0",
+  "author": {
+    "name": "David Pate",
+    "email": "me@davidtpate.com",
+    "url": "http://davidtpate.com"
+  },
+  "keywords": [
+    "uri",
+    "url",
+    "type",
+    "checker",
+    "validator",
+    "validate"
+  ],
+  "homepage": "https://github.com/DavidTPate/isuri",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/DavidTPate/isuri.git"
+  },
+  "bugs": {
+    "url": "https://github.com/DavidTPate/isuri/issues"
+  },
+  "license": "MIT",
+  "main": "index.js",
+  "devDependencies": {
+    "beautify-benchmark": "^0.x",
+    "benchmark": "^1.x",
+    "grunt": "^0.4.5",
+    "grunt-bump": "^0.3.1",
+    "grunt-contrib-clean": "^0.6.0",
+    "grunt-contrib-copy": "^0.8.0",
+    "grunt-contrib-uglify": "^0.9.1",
+    "istanbul": "^0.x",
+    "load-grunt-config": "^0.17.2",
+    "mocha": "^1.x",
+    "should": "^4.x"
+  },
+  "scripts": {
+    "test": "mocha --check-leaks --bail --reporter spec test/",
+    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --check-leaks --reporter dot test/",
+    "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --check-leaks --reporter spec test/"
+  }
 }
